refactor: import error from the correct place
