+++ lib/errors/index.ts
@@ -1,2 +1,3 @@
 export * from './ExecuteMigrationError';
 export * from './DbConnectionError';
+export * from './ConfigFileNotFoundError';
